+++ test/td_dd/data_structures/search_test.exs
@@ -5,6 +5,8 @@ defmodule TdDd.DataStructures.SearchTest do
 
   alias TdDd.DataStructures.Search
 
+  @moduletag sandbox: :shared
+
   @aggregations %{
     "foo" => %{
       "buckets" => [%{"key" => "bar"}, %{"key" => "baz"}]
@@ -69,6 +71,152 @@ defmodule TdDd.DataStructures.SearchTest do
         assert {:ok, %{"foo" => %{values: ["bar", "baz"]}}} =
                  Search.get_filter_values(claims, permission, %{})
       end
+    end
+  end
+
+  describe "vector/3" do
+    setup do
+      start_supervised!(TdDd.Search.StructureEnricher)
+      :ok
+    end
+
+    @tag authentication: [role: "admin"]
+    test "puts vector search in elastic", %{claims: claims} do
+      expect(ElasticsearchMock, :request, fn _, :post, "/structures/_search", request, _ ->
+        assert request == %{
+                 sort: ["_score"],
+                 _source: %{excludes: ["embeddings"]},
+                 knn: %{
+                   "field" => "embeddings.vector_foo",
+                   "filter" => %{bool: %{"filter" => %{match_all: %{}}}},
+                   "k" => 10,
+                   "num_candidates" => 200,
+                   "query_vector" => [54.0, 10.2, -2.0],
+                   "similarity" => 0.5
+                 }
+               }
+
+        SearchHelpers.hits_response([insert(:data_structure_version)])
+      end)
+
+      params = %{
+        "field" => "embeddings.vector_foo",
+        "k" => 10,
+        "similarity" => 0.5,
+        "num_candidates" => 200,
+        "query_vector" => [54.0, 10.2, -2.0]
+      }
+
+      assert %{total: 1, results: [_result]} =
+               Search.vector(claims, :view_data_structure, params)
+    end
+
+    @tag authentication: [role: "admin"]
+    test "includes similarity in response", %{claims: claims} do
+      expect(ElasticsearchMock, :request, fn _, :post, "/structures/_search", request, _ ->
+        assert request == %{
+                 sort: ["_score"],
+                 _source: %{excludes: ["embeddings"]},
+                 knn: %{
+                   "field" => "embeddings.vector_foo",
+                   "filter" => %{bool: %{"filter" => %{match_all: %{}}}},
+                   "k" => 10,
+                   "num_candidates" => 200,
+                   "query_vector" => [54.0, 10.2, -2.0],
+                   "similarity" => 0.5
+                 }
+               }
+
+        SearchHelpers.hits_response([insert(:data_structure_version)])
+      end)
+
+      params = %{
+        "field" => "embeddings.vector_foo",
+        "k" => 10,
+        "similarity" => 0.5,
+        "num_candidates" => 200,
+        "query_vector" => [54.0, 10.2, -2.0]
+      }
+
+      assert %{total: 1, results: [result]} =
+               Search.vector(claims, :view_data_structure, params, similarity: :cosine)
+
+      assert result.similarity == 1.0
+    end
+
+    @tag authentication: [role: "admin"]
+    test "exludes structure ids from search", %{claims: claims} do
+      expect(ElasticsearchMock, :request, fn _, :post, "/structures/_search", request, _ ->
+        assert request == %{
+                 sort: ["_score"],
+                 _source: %{excludes: ["embeddings"]},
+                 knn: %{
+                   "field" => "embeddings.vector_foo",
+                   "filter" => %{
+                     bool: %{
+                       "filter" => %{match_all: %{}},
+                       "must_not" => [%{term: %{"data_structure_id" => "1"}}]
+                     }
+                   },
+                   "k" => 10,
+                   "num_candidates" => 200,
+                   "query_vector" => [54.0, 10.2, -2.0],
+                   "similarity" => 0.5
+                 }
+               }
+
+        SearchHelpers.hits_response([insert(:data_structure_version)])
+      end)
+
+      params = %{
+        "field" => "embeddings.vector_foo",
+        "k" => 10,
+        "similarity" => 0.5,
+        "num_candidates" => 200,
+        "query_vector" => [54.0, 10.2, -2.0],
+        "structure_ids" => ["1"]
+      }
+
+      assert %{total: 1, results: [_result]} =
+               Search.vector(claims, :view_data_structure, params)
+    end
+
+    @tag authentication: [role: "user", permissions: ["view_data_structure"]]
+    test "puts domain filters in vector search", %{claims: claims, domain: domain} do
+      expect(ElasticsearchMock, :request, fn _, :post, "/structures/_search", request, _ ->
+        assert request == %{
+                 sort: ["_score"],
+                 _source: %{excludes: ["embeddings"]},
+                 knn: %{
+                   "field" => "embeddings.vector_foo",
+                   "filter" => %{
+                     bool: %{
+                       "filter" => [
+                         %{term: %{"domain_ids" => domain.id}},
+                         %{term: %{"confidential" => false}}
+                       ]
+                     }
+                   },
+                   "k" => 10,
+                   "num_candidates" => 200,
+                   "query_vector" => [54.0, 10.2, -2.0],
+                   "similarity" => 0.5
+                 }
+               }
+
+        SearchHelpers.hits_response([insert(:data_structure_version)])
+      end)
+
+      params = %{
+        "field" => "embeddings.vector_foo",
+        "k" => 10,
+        "similarity" => 0.5,
+        "num_candidates" => 200,
+        "query_vector" => [54.0, 10.2, -2.0]
+      }
+
+      assert %{total: 1, results: [_result]} =
+               Search.vector(claims, :view_data_structure, params)
     end
   end
 end